Add casts / change types to suppress warnings

+++ source/core/camera/pinhole_camera.h
@@ -21,7 +21,7 @@ namespace Lantern {
 class PinholeCamera {
 public:
 	PinholeCamera();
-	PinholeCamera(float phi, float theta, float radius, float clientWidth, float clientHeight, float3 target = float3(0.0f), float fov = M_PI_2, ReconstructionFilter::Type filterType = ReconstructionFilter::Type::Tent);
+	PinholeCamera(float phi, float theta, float radius, uint clientWidth, uint clientHeight, float3 target = float3(0.0f), float fov = M_PI_2, ReconstructionFilter::Type filterType = ReconstructionFilter::Type::Tent);
 	
 public:
 	// Frame Data
@@ -98,9 +98,9 @@ private:
 	* @return    The position of the camera
 	*/
 	void UpdateOrigin() {
-		float x = m_radius * std::sinf(m_phi) * std::sinf(m_theta);
-		float y = m_radius * std::cosf(m_phi);
-		float z = m_radius * std::sinf(m_phi) * std::cosf(m_theta);
+		float x = m_radius * sinf(m_phi) * sinf(m_theta);
+		float y = m_radius * cosf(m_phi);
+		float z = m_radius * sinf(m_phi) * cosf(m_theta);
 
 		m_origin = m_target + float3(x, y, z);
 	}
