+++ source/core/materials/media/isotropic_scattering_medium.h
@@ -29,7 +29,7 @@ private:
 
 public:
 	float SampleDistance(UniformSampler *sampler, float tFar, float *weight, float *pdf) const override {
-		float distance = -std::logf(sampler->NextFloat()) / m_scatteringCoefficient;
+		float distance = -logf(sampler->NextFloat()) / m_scatteringCoefficient;
 		if (distance >= tFar) {
 			*pdf = 1.0f;
 			return tFar;
@@ -39,12 +39,12 @@ public:
 		return distance;
 	}
 	float3a SampleScatterDirection(UniformSampler *sampler, float3a &wo, float *pdf) const override {
-		*pdf = 0.25f * M_1_PI; // 1 / (4 * PI)
+		*pdf = 0.25f * (float)M_1_PI; // 1 / (4 * PI)
 		return UniformSampleSphere(sampler);
 	}
 
 	float ScatterDirectionPdf(float3a &wi, float3a &wo) const override {
-		return 0.25f * M_1_PI; // 1 / (4 * PI)
+		return 0.25f * (float)M_1_PI; // 1 / (4 * PI)
 	}
 	float3 Transmission(float distance) const override {
 		return exp(-m_absorptionCoefficient * distance);
